+++ src/test/java/org/eclipse/keyple/core/service/LocalReaderAdapterTest.java
@@ -362,8 +362,6 @@ public class LocalReaderAdapterTest {
     assertThat(localReaderAdapter.isLogicalChannelOpen()).isFalse();
   }
 
-  // todo: selectByAid with AutonomousSelectionReaderSpi
-
   /*
    * Transmit card request
    */